refactored structure of color components
