+++ src/RGBlent.js
@@ -2,11 +2,11 @@ import React from "react";
 import styled from "styled-components";
 import { Container, Col, Row, Hr } from "@bootstrap-styled/v4";
 import { AuthProvider } from "./components/AuthProvider.js";
-import { ColorProvider } from "./components/ColorProvider.js";
+import { ColorProvider } from "./components/color/ColorProvider.js";
 import { PaletteProvider } from "./components/palette/PaletteProvider.js";
 import { Palette } from "./components/palette/Palette.js";
-import { Picker } from "./components/Picker.js";
-import { Detail } from "./components/Detail.js";
+import { Picker } from "./components/color/Picker.js";
+import { Detail } from "./components/color/Detail.js";
 import { AuthForm } from "./components/AuthForm.js";
 import { NameWindow } from "./components/NameWindow.js";
